refactor ui trace graph layout helpers
